+++ bindings/java/native/src/classes/client/mqtt.rs
@@ -1,61 +1,94 @@
-// Copyright 2021 IOTA Stiftung
+// Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
-// use crate::client::Client;
-// use pyo3::{conversion::ToPyObject, exceptions, prelude::*, types::PyTuple};
-// use std::convert::TryInto;
-//
-// #[pymethods]
-// impl Client {
-// fn subscribe_topic(&mut self, topic: &str, callback: PyObject) -> PyResult<()> {
-// self.subscribe_topics([topic].to_vec(), callback)
-// }
-//
-// fn subscribe_topics(&mut self, topics: Vec<&str>, callback: PyObject) -> PyResult<()> {
-// let result = crate::block_on(
-// self.client
-// .subscriber()
-// .with_topics(
-// topics
-// .iter()
-// .map(|&topic| topic.to_owned()[..].try_into().unwrap())
-// .collect(),
-// )
-// .subscribe(move |event| {
-// We need to clone it because `callback` has type `Py<PyAny>`,
-// which does not implement the `Copy` trait
-// let callback_copy = callback.clone();
-// Pare the event_string first to reduce the gil blocking period
-// let event_string = serde_json::to_string(&event).unwrap();
-// crate::spawn_blocking(move || {
-// let gil = Python::acquire_gil();
-// let py = gil.python();
-// let args = PyTuple::new(py, &[event_string]);
-// callback_copy.call1(py, args).unwrap_or_else(|_| {
-// PyErr::new::<exceptions::PyTypeError, _>("Unable to use the python callback function!")
-// .to_object(py)
-// })
-// });
-// }),
-// );
-// match result {
-// Err(err) => Err(PyErr::new::<exceptions::PyTypeError, _>(err.to_string())),
-// Ok(()) => Ok(()),
-// }
-// }
-//
-// fn unsubscribe(&mut self) -> PyResult<()> {
-// let result = crate::block_on(self.client.subscriber().unsubscribe());
-// match result {
-// Err(err) => Err(PyErr::new::<exceptions::PyTypeError, _>(err.to_string())),
-// Ok(()) => Ok(()),
-// }
-// }
-//
-// fn disconnect(&mut self) -> PyResult<()> {
-// let result = crate::block_on(self.client.subscriber().disconnect());
-// match result {
-// Err(err) => Err(PyErr::new::<exceptions::PyTypeError, _>(err.to_string())),
-// Ok(()) => Ok(()),
-// }
-// }
-// }
+use std::{cell::RefCell, rc::Rc};
+
+use iota_client::{
+    node::{MqttManager as RustMqttManager, MqttTopicManager as RustMqttTopicManager, Topic},
+    TopicEvent,
+};
+
+use crate::{full_node_api::Client, Result};
+
+use anyhow::anyhow;
+
+pub trait MqttListener {
+    fn on_event(&self, event: TopicEvent);
+}
+
+pub struct MqttManager<'a> {
+    client: &'a mut Client,
+}
+
+impl<'a> MqttManager<'a> {
+    pub fn new(client: &'a mut Client) -> Self {
+        Self { client }
+    }
+
+    pub fn with_topic(&mut self, topic: Topic) -> MqttTopicManager {
+        MqttTopicManager::new(RustMqttManager::new(self.client.borrow_mut()).with_topic(topic))
+    }
+
+    pub fn with_topics(&mut self, topics: Vec<Topic>) -> MqttTopicManager {
+        MqttTopicManager::new(RustMqttManager::new(self.client.borrow_mut()).with_topics(topics))
+    }
+
+    pub fn unsubscribe(&mut self) -> Result<()> {
+        let res = crate::block_on(async { RustMqttManager::new(self.client.borrow_mut()).unsubscribe().await });
+        match res {
+            Ok(()) => Ok(()),
+            Err(e) => Err(anyhow!(e.to_string())),
+        }
+    }
+
+    pub fn disconnect(&mut self) -> Result<()> {
+        let res = crate::block_on(async { RustMqttManager::new(self.client.borrow_mut()).disconnect().await });
+        match res {
+            Ok(()) => Ok(()),
+            Err(e) => Err(anyhow!(e.to_string())),
+        }
+    }
+}
+
+pub struct MqttTopicManager<'a>(Rc<RefCell<Option<RustMqttTopicManager<'a>>>>);
+
+impl<'a> MqttTopicManager<'a> {
+    pub fn new(manager: RustMqttTopicManager<'a>) -> Self {
+        Self(Rc::new(RefCell::new(Option::from(manager))))
+    }
+
+    pub fn with_topic(&self, topic: Topic) -> Self {
+        let new_manager = self.0.borrow_mut().take().unwrap().with_topic(topic);
+        MqttTopicManager::new(new_manager)
+    }
+
+    pub fn with_topics(&self, topics: Vec<Topic>) -> Self {
+        let new_manager = self.0.borrow_mut().take().unwrap().with_topics(topics);
+        MqttTopicManager::new(new_manager)
+    }
+
+    pub fn subscribe(&self, cb: Box<dyn MqttListener + Send + Sync + 'static>) -> Result<()> {
+        let new_manager = self.0.borrow_mut().take().unwrap();
+
+        let res = crate::block_on(async {
+            new_manager
+                .subscribe(move |event| {
+                    cb.on_event(event.clone());
+                })
+                .await
+        });
+
+        match res {
+            Ok(()) => Ok(()),
+            Err(e) => Err(anyhow!(e.to_string())),
+        }
+    }
+
+    pub fn unsubscribe(&self) -> Result<()> {
+        let new_manager = self.0.borrow_mut().take().unwrap();
+        let res = crate::block_on(async move { new_manager.unsubscribe().await });
+        match res {
+            Ok(()) => Ok(()),
+            Err(e) => Err(anyhow!(e.to_string())),
+        }
+    }
+}
